+++ src/colorManager/colorManager.h
@@ -36,7 +36,7 @@ void colorMngrSetSingle(Color_t *color, ZephyrRgbPixel_t *pixels,
  * @param pixels      The pixel buffer.
  * @param pixelCnt    The count of pixel to manage.
  */
-void colorMngrApplyFade(uint32_t fadeLvl, ZephyrRgbPixel_t *pixels,
+void colorMngrApplyFade(uint8_t fadeLvl, ZephyrRgbPixel_t *pixels,
                         size_t pixelCnt);
 
 /**
@@ -48,7 +48,7 @@ void colorMngrApplyFade(uint32_t fadeLvl, ZephyrRgbPixel_t *pixels,
  * @param pixels      The pixel buffer.
  * @param pixelCnt    The count pixel to manage.
  */
-void colorMngrApplyFadeTrail(uint32_t fadeLvl, uint32_t fadeStart,
+void colorMngrApplyFadeTrail(uint8_t fadeLvl, uint32_t fadeStart,
                              bool isAscending, ZephyrRgbPixel_t *pixels,
                              size_t pixelCnt);
